+++ packages/cli/src/commands/sandbox/sandbox_command.test.ts
@@ -20,6 +20,7 @@ import {
 import { createSandboxSecretCommand } from './sandbox-secret/sandbox_secret_command_factory.js';
 import { ClientConfigGeneratorAdapter } from '../../client-config/client_config_generator_adapter.js';
 import { CommandMiddleware } from '../../command_middleware.js';
+import { PackageManagerController } from '@aws-amplify/plugin-types';
 
 mock.method(fsp, 'mkdir', () => Promise.resolve());
 
@@ -52,6 +53,11 @@ void describe('sandbox command', () => {
   );
   const sandboxProfile = 'test-sandbox';
 
+  const allowsSignalPropagationMock = mock.fn(() => true);
+  const packageManagerControllerMock = {
+    allowsSignalPropagation: allowsSignalPropagationMock,
+  } as unknown as PackageManagerController;
+
   beforeEach(async () => {
     const sandboxFactory = new SandboxSingletonFactory(
       () =>
@@ -73,6 +79,7 @@ void describe('sandbox command', () => {
       [sandboxDeleteCommand, createSandboxSecretCommand()],
       clientConfigGeneratorAdapterMock,
       commandMiddleware,
+      packageManagerControllerMock,
       () => ({
         successfulDeployment: [clientConfigGenerationMock],
         successfulDeletion: [clientConfigDeletionMock],
@@ -232,6 +239,39 @@ void describe('sandbox command', () => {
     assert.equal(sandboxDeleteMock.mock.callCount(), 0);
   });
 
+  void it('Does not prompt for deleting the sandbox if package manager does not allow signal propagation', async (contextual) => {
+    allowsSignalPropagationMock.mock.mockImplementationOnce(() => false);
+
+    // Mock process and extract the sigint handler after calling the sandbox command
+    const processSignal = contextual.mock.method(process, 'on', () => {
+      /* no op */
+    });
+    const sandboxStartMock = contextual.mock.method(
+      sandbox,
+      'start',
+      async () => Promise.resolve()
+    );
+
+    const printerMock = contextual.mock.method(printer, 'print', () => {});
+
+    await commandRunner.runCommand('sandbox');
+
+    // Similar to the previous test's 0ms timeout. Without this tests in github action are failing
+    // but working locally
+    await new Promise((resolve) => setTimeout(resolve, 0));
+    const sigIntHandlerFn = processSignal.mock.calls[0].arguments[1];
+    if (sigIntHandlerFn) sigIntHandlerFn();
+
+    assert.equal(sandboxStartMock.mock.callCount(), 1);
+    assert.equal(printerMock.mock.callCount(), 1);
+    assert.equal(
+      printerMock.mock.calls[0].arguments[0],
+      `Stopping the sandbox process. To delete the sandbox, run ${format.normalizeAmpxCommand(
+        'sandbox delete'
+      )}`
+    );
+  });
+
   void it('starts sandbox with user provided invalid AWS profile', async () => {
     const profileErr = new Error('some profile error');
     mockHandleProfile.mock.mockImplementationOnce(() => {
@@ -272,6 +312,7 @@ void describe('sandbox command', () => {
       [],
       clientConfigGeneratorAdapterMock,
       commandMiddleware,
+      packageManagerControllerMock,
       undefined
     );
     const parser = yargs().command(sandboxCommand as unknown as CommandModule);