+++ packages/backend-deployer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/backend-deployer
 
+## 1.0.2
+
+### Patch Changes
+
+- 697bc8a: Prevent CTRL-C handling for yarn classic package manager
+- Updated dependencies [697bc8a]
+- Updated dependencies [c784e40]
+  - @aws-amplify/plugin-types@1.1.0
+  - @aws-amplify/platform-core@1.0.3
+
 ## 1.0.1
 
 ### Patch Changes